+++ compiler-examples/ck-compiler/src/test/java/com/arjunsk/compiler/ck/parser/ParserTest.java
@@ -29,6 +29,8 @@ public class ParserTest {
     Assert.assertEquals(1, tree.getChild(2).getChildCount()); // Statement
     Assert.assertEquals(1, tree.getChild(2).getChild(0).getChildCount()); // SHOW show 23
 
-    System.out.println(tree.toStringTree());
+    Assert.assertEquals(
+        "( ProgramContext( ( StatementContext( ( LetContext( a  10 ) ) ) )  ( StatementContext( ( ShowContext( a ) ) ) )  ( StatementContext( ( ShowContext( 20 ) ) ) ) ) )",
+        tree.toStringTree());
   }
 }
