+++ compiler-examples/ck-compiler/src/test/java/com/arjunsk/compiler/utils/FileReaderUtil.java
@@ -1,4 +1,4 @@
-package com.arjunsk.compiler.ck.utils;
+package com.arjunsk.compiler.utils;
 
 import java.io.BufferedReader;
 import java.io.InputStream;
